backend/requirements.txt
fastapi
uvicorn
pydantic
sqlmodel
passlib[bcrypt] 
password-validator 
email-validator 
pyjwt 